Show sub proposals on custom domain
