+++ src/views/SetupView.vue
@@ -36,13 +36,13 @@ const creatingSpace = ref(false);
 const { t } = useI18n();
 const { pendingENSRecord, uriAddress, loadUriAddress } = useSpaceController();
 const { send } = useClient();
-const { loadExtentedSpaces, extentedSpaces } = useExtendedSpaces();
+const { loadExtendedSpaces, extendedSpaces } = useExtendedSpaces();
 
 const currentStep = computed(() => Number(route.query.step));
 
 async function checkIfSpaceExists() {
-  await loadExtentedSpaces([route.params.ens as string]);
-  if (extentedSpaces.value?.some(space => space.id === route.params.ens)) {
+  await loadExtendedSpaces([route.params.ens as string]);
+  if (extendedSpaces.value?.some(space => space.id === route.params.ens)) {
     return;
   } else {
     await sleep(5000);
